Struct iron_hmac::Hmac256Authentication

pub struct Hmac256Authentication {
    // some fields omitted
}

Iron middleware for validation hmac headers on requests and signing responses.

The algorithm employed is as follows.

hmac(secret, hmac(secret, method) + hmac(secret, path) + hmac(secret, body))

Methods

impl Hmac256Authentication

fn middleware<K: Into<SecretKey>, S: Into<String>>(secret: K, hmac_header_key: S) -> (Hmac256Authentication, Hmac256Authentication)

Build Hmac256Authentication before and after middleware given a secret and header key.